+++ RELEASE-NOTES.md
@@ -13,6 +13,7 @@
 - Decentralize the menus.
 - Move all configurable properties from config.properties to APPLICATION.JNLP that is used by BCL.
 - Register all pop-up dialogs for idle monitoring.
+- Improve the logging configuration.
 
 ## v1.1.2